% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{qual_catalogue}
\alias{qual_catalogue}
\title{Summarise references as a data frame}
\usage{
qual_catalogue(refs)
}
\arguments{
\item{refs}{A named list of references.}
}
\value{
A data frame with one row per reference.
}
\description{
Summarise references as a data frame
}
