% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{qual_reference_id}
\alias{qual_reference_id}
\title{Unique, thread-tagged id for a reference}
\usage{
qual_reference_id(ref)
}
\arguments{
\item{ref}{A reference list.}
}
\value{
\code{"<model>__<method>__t<threads>"}.
}
\description{
Unique, thread-tagged id for a reference
}
