% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{qual_reference_validate}
\alias{qual_reference_validate}
\title{Validate a reference has the required structure}
\usage{
qual_reference_validate(ref)
}
\arguments{
\item{ref}{A reference list.}
}
\value{
\code{TRUE} invisibly if valid; errors otherwise.
}
\description{
Validate a reference has the required structure
}
