% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-io.R
\name{qual_read_dataset}
\alias{qual_read_dataset}
\title{Load a frozen dataset by file reference}
\usage{
qual_read_dataset(ref)
}
\arguments{
\item{ref}{The dataset file name under inst/extdata/ (or an absolute path).}
}
\value{
A data frame.
}
\description{
Load a frozen dataset by file reference
}
